- let version be %{main_version}.%{daily_version}
Changed files:
    clamav-database.spec -> 1.4

+++ clamav-database.spec
@@ -1,9 +1,12 @@
+%define		main_version		41
+%define		daily_version		2384
 %define		database_version	20061227
+%define		_rel	1
 Summary:	Virus database for clamav
 Summary(pl):	Bazy wirusów dla clamava
 Name:		clamav-database
-Version:	0.88.7.%{database_version}
-Release:	1
+Version:	%{main_version}.%{daily_version}
+Release:	%{database_version}.%{_rel}
 License:	GPL
 Group:		Applications/Databases
 Source0:	http://db.local.clamav.net/daily.cvd
@@ -11,6 +14,7 @@ Source0:	http://db.local.clamav.net/daily.cvd
 Source1:	http://db.local.clamav.net/main.cvd
 # Source1-md5:	347c99544205184fbc1bd23fd7cfd782
 URL:		http://www.clamav.net/
+BuildRequires:	file
 Requires:	clamav
 BuildArch:	noarch
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
@@ -23,13 +27,24 @@ Bazy wirusów dla clamava (aktualizowana %{database_version}).
 
 %prep
 %setup -qcT
+cp -a %{SOURCE0} %{SOURCE1} .
+
+%build
+main_version=$(file main.cvd | awk -F, '/version/{print $2}' | awk '{print $NF}')
+daily_version=$(file daily.cvd | awk -F, '/version/{print $2}' | awk '{print $NF}')
+if [ "$main_version" != %{main_version} ]; then
+	: Update %%define main_version $main_version, and retry
+	exit 1
+fi
+if [ "$daily_version" != %{daily_version} ]; then
+	: Update %%define daily_version $daily_version, and retry
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
-install -d $RPM_BUILD_ROOT{/var/lib/clamav,%{_sbindir}}
-
-install %{SOURCE0} $RPM_BUILD_ROOT/var/lib/clamav
-install %{SOURCE1} $RPM_BUILD_ROOT/var/lib/clamav
+install -d $RPM_BUILD_ROOT/var/lib/clamav
+install *.cvd $RPM_BUILD_ROOT/var/lib/clamav
 
 %clean
 rm -rf $RPM_BUILD_ROOT
